ingest/googlesheets_bq/requirements.txt
gspread==6.2.1
google-auth==2.40.3
google-cloud-bigquery==3.34.0
functions-framework==3.8.3